+++ example/src/App.tsx
@@ -1,28 +1,37 @@
-import { ChangeEvent, useEffect, useRef, useState } from "react";
+import { ChangeEvent, forwardRef, useEffect, useRef, useState } from "react";
 import { Input, TheMaskInputProps } from "../../src";
 
 type Value = string | number;
 
-const StyledInput = (props: TheMaskInputProps) => (
+const StyledInput = forwardRef((props: TheMaskInputProps, ref: any) => (
 	<fieldset className="flex flex-col gap-2">
 		<label className="text-white uppercase text-xs">{props.placeholder}</label>
 		<Input
 			{...props}
+			ref={ref}
 			name={props.name ?? (props.mask as string)}
 			required
 			className="p-1 border border-slate-300 text-white antialiased rounded"
 		/>
 	</fieldset>
-);
+));
 
 export const Controlled = () => {
+	const ref = useRef<HTMLInputElement>(null);
 	const [value, setValue] = useState("11122233344");
+	useEffect(() => {
+		if (ref.current) {
+			ref.current.focus();
+		}
+	}, []);
 	return (
 		<fieldset className="p-4 border border-slate-400">
 			{value}
 			<StyledInput
 				mask="cpf"
 				value={value}
+				ref={ref}
+				name="random"
 				onChange={(e) => {
 					setValue(e.target.value);
 				}}
@@ -37,7 +46,8 @@ export const Controlled = () => {
 				}
 			>
 				Change
-			</button></fieldset>
+			</button>
+		</fieldset>
 	);
 };
 
@@ -97,66 +107,66 @@ export default function App() {
 	return (
 		<div className="p-8 flex flex-wrap gap-8">
 			<Controlled />
-			{/*<InputsWithDefault />*/}
-			{/*{show && <span id="test">Success</span>}*/}
-			{/*<form*/}
-			{/*	className="flex gap-6 flex-wrap items-end"*/}
-			{/*	onSubmit={(e) => {*/}
-			{/*		e.preventDefault();*/}
-			{/*		setShow(true);*/}
-			{/*	}}*/}
-			{/*>*/}
-			{/*	<StyledInput value={state.int} onChange={onChange} mask="int" name="int" type="number" placeholder="Integer" />*/}
-			{/*	<StyledInput value={state.percent} onChange={onChange} mask="percent" name="percent" placeholder="percent" />*/}
-			{/*	<StyledInput value={state.cellTelephone} onChange={onChange} mask="cellTelephone" name="cellTelephone" placeholder="cellTelephone" />*/}
-			{/*	<StyledInput value={state.cellphone} onChange={onChange} mask="cellphone" name="cellphone" placeholder="cellphone" />*/}
-			{/*	<StyledInput value={state.cep} onChange={onChange} mask="cep" name="cep" placeholder="cep" />*/}
-			{/*	<StyledInput value={state.cnpj} onChange={onChange} mask="cnpj" name="cnpj" placeholder="cnpj" />*/}
-			{/*	<StyledInput value={state.cpfCnpj} onChange={onChange} mask="cpfCnpj" name="cpfCnpj" placeholder="cpfCnpj" />*/}
-			{/*	<StyledInput value={state.color} onChange={onChange} mask="color" name="color" placeholder="color" />*/}
-			{/*	<StyledInput value={state.cpf} onChange={onChange} mask="cpf" name="cpf" placeholder="cpf" />*/}
-			{/*	<StyledInput value={state.creditCard} onChange={onChange} mask="creditCard" name="creditCard" placeholder="creditCard" />*/}
-			{/*	<StyledInput value={state.date} onChange={onChange} mask="date" name="date" placeholder="date" />*/}
-			{/*	<StyledInput value={state.isoDate} onChange={onChange} mask="isoDate" name="isoDate" placeholder="isoDate" />*/}
-			{/*	<StyledInput value={state.money} onChange={onChange} mask="money" name="money" placeholder="Money - asNumber" />*/}
-			{/*	<StyledInput*/}
-			{/*		value={state.Currency}*/}
-			{/*		onChange={onChange}*/}
-			{/*		name="Currency"*/}
-			{/*		placeholder="Currency - Symbol"*/}
-			{/*		mask="currency"*/}
-			{/*		currency="USD"*/}
-			{/*		locale="en-US"*/}
-			{/*	/>*/}
-			{/*	<StyledInput*/}
-			{/*		value={state.CurrencyCodeUsd}*/}
-			{/*		onChange={onChange}*/}
-			{/*		name="CurrencyCodeUsd"*/}
-			{/*		placeholder="Currency - Code"*/}
-			{/*		mask="currency"*/}
-			{/*		currency="USD"*/}
-			{/*		currencyDisplay="code"*/}
-			{/*		locale="en-US"*/}
-			{/*	/>*/}
-			{/*	<StyledInput value={state.CurrencyCode} onChange={onChange} name="CurrencyCode" placeholder="Currency - Code" mask="currency" />*/}
-			{/*	<StyledInput value={state.telephone} onChange={onChange} mask="telephone" name="telephone" placeholder="telephone" />*/}
-			{/*	<StyledInput value={state.time} onChange={onChange} mask="time" name="time" placeholder="time" />*/}
-			{/*	<StyledInput value={state.uuid} onChange={onChange} mask="uuid" name="uuid" placeholder="uuid" />*/}
-			{/*	<StyledInput*/}
-			{/*		value={state.expiresIn}*/}
-			{/*		onChange={onChange}*/}
-			{/*		mask={[/\d/, /\d/, "/", /\d/, /\d/, /\d/, /\d/]}*/}
-			{/*		name="expiresIn"*/}
-			{/*		placeholder="expiresIn"*/}
-			{/*	/>*/}
-			{/*	<Input />*/}
-			{/*	<button id="submit" className="px-4 py-1 bg-blue-500 text-white rounded-lg h-fit" type="reset">*/}
-			{/*		Reset*/}
-			{/*	</button>*/}
-			{/*	<button id="submit" className="px-4 py-1 bg-blue-500 text-white rounded-lg h-fit" type="submit">*/}
-			{/*		Test pattern*/}
-			{/*	</button>*/}
-			{/*</form>*/}
+			<InputsWithDefault />
+			{show && <span id="test">Success</span>}
+			<form
+				className="flex gap-6 flex-wrap items-end"
+				onSubmit={(e) => {
+					e.preventDefault();
+					setShow(true);
+				}}
+			>
+				<StyledInput value={state.int} onChange={onChange} mask="int" name="int" type="number" placeholder="Integer" />
+				<StyledInput value={state.percent} onChange={onChange} mask="percent" name="percent" placeholder="percent" />
+				<StyledInput value={state.cellTelephone} onChange={onChange} mask="cellTelephone" name="cellTelephone" placeholder="cellTelephone" />
+				<StyledInput value={state.cellphone} onChange={onChange} mask="cellphone" name="cellphone" placeholder="cellphone" />
+				<StyledInput value={state.cep} onChange={onChange} mask="cep" name="cep" placeholder="cep" />
+				<StyledInput value={state.cnpj} onChange={onChange} mask="cnpj" name="cnpj" placeholder="cnpj" />
+				<StyledInput value={state.cpfCnpj} onChange={onChange} mask="cpfCnpj" name="cpfCnpj" placeholder="cpfCnpj" />
+				<StyledInput value={state.color} onChange={onChange} mask="color" name="color" placeholder="color" />
+				<StyledInput value={state.cpf} onChange={onChange} mask="cpf" name="cpf" placeholder="cpf" />
+				<StyledInput value={state.creditCard} onChange={onChange} mask="creditCard" name="creditCard" placeholder="creditCard" />
+				<StyledInput value={state.date} onChange={onChange} mask="date" name="date" placeholder="date" />
+				<StyledInput value={state.isoDate} onChange={onChange} mask="isoDate" name="isoDate" placeholder="isoDate" />
+				<StyledInput value={state.money} onChange={onChange} mask="money" name="money" placeholder="Money - asNumber" />
+				<StyledInput
+					value={state.Currency}
+					onChange={onChange}
+					name="Currency"
+					placeholder="Currency - Symbol"
+					mask="currency"
+					currency="USD"
+					locale="en-US"
+				/>
+				<StyledInput
+					value={state.CurrencyCodeUsd}
+					onChange={onChange}
+					name="CurrencyCodeUsd"
+					placeholder="Currency - Code"
+					mask="currency"
+					currency="USD"
+					currencyDisplay="code"
+					locale="en-US"
+				/>
+				<StyledInput value={state.CurrencyCode} onChange={onChange} name="CurrencyCode" placeholder="Currency - Code" mask="currency" />
+				<StyledInput value={state.telephone} onChange={onChange} mask="telephone" name="telephone" placeholder="telephone" />
+				<StyledInput value={state.time} onChange={onChange} mask="time" name="time" placeholder="time" />
+				<StyledInput value={state.uuid} onChange={onChange} mask="uuid" name="uuid" placeholder="uuid" />
+				<StyledInput
+					value={state.expiresIn}
+					onChange={onChange}
+					mask={[/\d/, /\d/, "/", /\d/, /\d/, /\d/, /\d/]}
+					name="expiresIn"
+					placeholder="expiresIn"
+				/>
+				<Input />
+				<button id="submit" className="px-4 py-1 bg-blue-500 text-white rounded-lg h-fit" type="reset">
+					Reset
+				</button>
+				<button id="submit" className="px-4 py-1 bg-blue-500 text-white rounded-lg h-fit" type="submit">
+					Test pattern
+				</button>
+			</form>
 		</div>
 	);
 }
