--- ssr-example/next.config.js
@@ -1,4 +1,0 @@
-/** @type {import('next').NextConfig} */
-const nextConfig = {}
-
-module.exports = nextConfig
